refactor: split UnrealMCP C++ handler routes
