+++ Plugins/UnrealMCP/Source/UnrealMCP/Private/Commands/EpicUnrealMCPProceduralCommands.cpp
@@ -1,6 +1,23 @@
+// =====================================================================
+// EpicUnrealMCPProceduralCommands
+//
+// Phase 4 (Issue #31) trimmed this file to host only the procedural
+// generation surface and the single request_cognitive_processing
+// command.  Physics / Validation / Instance handlers were extracted into
+// dedicated classes:
+//   - FEpicUnrealMCPPhysicsCommands     (route 22)
+//   - FEpicUnrealMCPValidationCommands  (route 23)
+//   - FEpicUnrealMCPInstanceCommands    (route 24)
+//
+// Adding a new procedural command that fits the existing surface should
+// stay here; otherwise spin up a new FEpicUnrealMCPXxxCommands so this
+// file does not regrow back to the pre-Phase-4 size.
+// =====================================================================
+
 #include "Commands/EpicUnrealMCPProceduralCommands.h"
 #include "Commands/EpicUnrealMCPCommonUtils.h"
 #include "EpicUnrealMCPBridge.h"
+
 #include "Editor.h"
 #include "EditorViewportClient.h"
 #include "LevelEditorViewport.h"
@@ -15,116 +32,8 @@
 #include "Materials/MaterialInstanceDynamic.h"
 #include "EditorAssetLibrary.h"
 #include "ScopedTransaction.h"
-#include "PhysicalMaterials/PhysicalMaterial.h"
-#include "PhysicsEngine/RadialForceComponent.h"
-#include "PhysicsEngine/RadialForceActor.h"
-#include "PhysicsEngine/PhysicsConstraintActor.h"
-#include "GameFramework/PhysicsVolume.h"
 #include "AssetRegistry/AssetRegistryModule.h"
 #include "AssetRegistry/IAssetRegistry.h"
-#include "Kismet2/KismetEditorUtilities.h"
-#include "Kismet2/CompilerResultsLog.h"
-#include "Engine/Blueprint.h"
-
-//   - guardrails: per-call max counts to keep accidental large requests safe
-
-// =====================================================================
-// Static file-local helpers (moved from EpicUnrealMCPEditorCommands.cpp
-// during the Phase 1 refactor).  Must be defined before first use below.
-// =====================================================================
-
-static AActor* FindDraftProxyActor(UWorld* World, const FString& ProxyName)
-{
-    if (!World) return nullptr;
-    FString McpIdTag = FString::Printf(TEXT("mcp_id:draft_%s"), *ProxyName);
-    for (TActorIterator<AActor> It(World); It; ++It)
-    {
-        if (It->Tags.Contains(FName(*McpIdTag)))
-        {
-            return *It;
-        }
-    }
-    return nullptr;
-}
-
-static UMaterialInterface* LoadDraftMaterial(UObject* Outer, bool bUseDither)
-{
-    if (!Outer) return nullptr;
-
-    // Try project-specific preset material first
-    UMaterialInterface* Mat = Cast<UMaterialInterface>(
-        UEditorAssetLibrary::LoadAsset(TEXT("/Game/Materials/M_DraftProxy"))
-    );
-    if (Mat) return Mat;
-
-    // Fallback: try dither variant
-    if (bUseDither)
-    {
-        Mat = Cast<UMaterialInterface>(
-            UEditorAssetLibrary::LoadAsset(TEXT("/Game/Materials/M_DraftProxy_Dither"))
-        );
-        if (Mat) return Mat;
-    }
-
-    // Final fallback: use the engine default translucent material
-    Mat = Cast<UMaterialInterface>(
-        UEditorAssetLibrary::LoadAsset(TEXT("/Engine/BasicShapes/BasicShapeMaterial"))
-    );
-    return Mat;
-}
-
-static UHierarchicalInstancedStaticMeshComponent* GetOrCreateHismComponent(AActor* Actor)
-{
-    if (!Actor) return nullptr;
-    UHierarchicalInstancedStaticMeshComponent* HISM = Actor->FindComponentByClass<UHierarchicalInstancedStaticMeshComponent>();
-    if (!HISM)
-    {
-        HISM = NewObject<UHierarchicalInstancedStaticMeshComponent>(Actor, UHierarchicalInstancedStaticMeshComponent::StaticClass(), TEXT("DraftHISM"));
-        HISM->RegisterComponent();
-        if (Actor->GetRootComponent())
-        {
-            HISM->AttachToComponent(Actor->GetRootComponent(), FAttachmentTransformRules::KeepRelativeTransform);
-        }
-        else
-        {
-            Actor->SetRootComponent(HISM);
-        }
-    }
-    return HISM;
-}
-
-static bool ParseInstanceEntry(const TSharedPtr<FJsonObject>& Obj, FVector& OutLocation, FVector& OutScale, FRotator& OutRotation)
-{
-    if (!Obj.IsValid()) return false;
-    const TSharedPtr<FJsonObject>* LocPtr = nullptr;
-    if (Obj->TryGetObjectField(TEXT("location"), LocPtr) && LocPtr)
-    {
-        double X = 0, Y = 0, Z = 0;
-        (*LocPtr)->TryGetNumberField(TEXT("x"), X);
-        (*LocPtr)->TryGetNumberField(TEXT("y"), Y);
-        (*LocPtr)->TryGetNumberField(TEXT("z"), Z);
-        OutLocation = FVector(X, Y, Z);
-    }
-    const TSharedPtr<FJsonObject>* ScalePtr = nullptr;
-    if (Obj->TryGetObjectField(TEXT("scale"), ScalePtr) && ScalePtr)
-    {
-        double X = 1, Y = 1, Z = 1;
-        (*ScalePtr)->TryGetNumberField(TEXT("x"), X);
-        (*ScalePtr)->TryGetNumberField(TEXT("y"), Y);
-        (*ScalePtr)->TryGetNumberField(TEXT("z"), Z);
-        OutScale = FVector(X, Y, Z);
-    }
-    const TSharedPtr<FJsonObject>* RotPtr = nullptr;
-    if (Obj->TryGetObjectField(TEXT("rotation"), RotPtr) && RotPtr)
-    {
-        double Pitch = 0, Yaw = 0, Roll = 0;
-        (*RotPtr)->TryGetNumberField(TEXT("pitch"), Pitch);
-        (*RotPtr)->TryGetNumberField(TEXT("yaw"), Yaw);
-        (*RotPtr)->TryGetNumberField(TEXT("roll"), Roll);
-        OutRotation = FRotator(Pitch, Yaw, Roll);
-    }
-    return true;
-}
 
 // =====================================================================
 // Class members
@@ -148,36 +57,14 @@ TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleCommand(const FS
     using Handler = TSharedPtr<FJsonObject>(FEpicUnrealMCPProceduralCommands::*)(const TSharedPtr<FJsonObject>&);
     static const TMap<FString, Handler> Dispatch = {
         // Procedural generation
-        {TEXT("spawn_tile_grid"), &FEpicUnrealMCPProceduralCommands::HandleSpawnTileGrid},
-        {TEXT("spawn_procedural_actor_batch"), &FEpicUnrealMCPProceduralCommands::HandleSpawnProceduralActorBatch},
+        {TEXT("spawn_tile_grid"),                  &FEpicUnrealMCPProceduralCommands::HandleSpawnTileGrid},
+        {TEXT("spawn_procedural_actor_batch"),     &FEpicUnrealMCPProceduralCommands::HandleSpawnProceduralActorBatch},
         {TEXT("create_spline_mesh_from_segments"), &FEpicUnrealMCPProceduralCommands::HandleCreateSplineMeshFromSegments},
         {TEXT("create_data_layer_for_generation"), &FEpicUnrealMCPProceduralCommands::HandleCreateDataLayerForGeneration},
-        {TEXT("clear_generated_group"), &FEpicUnrealMCPProceduralCommands::HandleClearGeneratedGroup},
+        {TEXT("clear_generated_group"),            &FEpicUnrealMCPProceduralCommands::HandleClearGeneratedGroup},
 
-        // Draft proxy (HISM visualization)
-        {TEXT("create_draft_proxy"), &FEpicUnrealMCPProceduralCommands::HandleCreateDraftProxy},
-        {TEXT("update_draft_proxy"), &FEpicUnrealMCPProceduralCommands::HandleUpdateDraftProxy},
-        {TEXT("delete_draft_proxy"), &FEpicUnrealMCPProceduralCommands::HandleDeleteDraftProxy},
-
-        // InstanceSet commands (HISM/ISM bulk instancing)
-        {TEXT("spawn_instance_set"), &FEpicUnrealMCPProceduralCommands::HandleSpawnInstanceSet},
-        {TEXT("update_instance_set"), &FEpicUnrealMCPProceduralCommands::HandleUpdateInstanceSet},
-        {TEXT("delete_instance_set"), &FEpicUnrealMCPProceduralCommands::HandleDeleteInstanceSet},
-        {TEXT("get_instance_set_state"), &FEpicUnrealMCPProceduralCommands::HandleGetInstanceSetState},
-        {TEXT("list_instance_sets"), &FEpicUnrealMCPProceduralCommands::HandleListInstanceSets},
-        {TEXT("request_cognitive_processing"), &FEpicUnrealMCPProceduralCommands::HandleRequestCognitiveProcessing},
-
-        // Physics commands (collision/physics/material/forces/constraints)
-        {TEXT("set_actor_collision_preset"), &FEpicUnrealMCPProceduralCommands::HandleSetActorCollisionPreset},
-        {TEXT("set_actor_physics"), &FEpicUnrealMCPProceduralCommands::HandleSetActorPhysics},
-        {TEXT("create_physical_material"), &FEpicUnrealMCPProceduralCommands::HandleCreatePhysicalMaterial},
-        {TEXT("spawn_radial_force"), &FEpicUnrealMCPProceduralCommands::HandleSpawnRadialForce},
-        {TEXT("spawn_physics_constraint"), &FEpicUnrealMCPProceduralCommands::HandleSpawnPhysicsConstraint},
-
-        // Validation commands
-        {TEXT("compile_all_blueprints"), &FEpicUnrealMCPProceduralCommands::HandleCompileAllBlueprints},
-        {TEXT("run_map_check"), &FEpicUnrealMCPProceduralCommands::HandleRunMapCheck},
-        {TEXT("find_broken_references"), &FEpicUnrealMCPProceduralCommands::HandleFindBrokenReferences},
+        // Cognitive processing (single-command surface kept here for now)
+        {TEXT("request_cognitive_processing"),     &FEpicUnrealMCPProceduralCommands::HandleRequestCognitiveProcessing},
     };
 
     const Handler* H = Dispatch.Find(CommandType);
@@ -186,1161 +73,8 @@ TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleCommand(const FS
         return (this->*(*H))(Params);
     }
 
-    return FEpicUnrealMCPCommonUtils::CreateErrorResponse(FString::Printf(TEXT("Unknown procedural command: %s"), *CommandType));
-}
-TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleCreateDraftProxy(const TSharedPtr<FJsonObject>& Params)
-{
-    FString ProxyName;
-    if (!Params->TryGetStringField(TEXT("proxy_name"), ProxyName) || ProxyName.IsEmpty())
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Missing 'proxy_name' parameter"));
-    }
-
-    FString MeshPath = TEXT("/Engine/BasicShapes/Cube.Cube");
-    Params->TryGetStringField(TEXT("mesh_path"), MeshPath);
-
-    FString MaterialPath;
-    Params->TryGetStringField(TEXT("material_path"), MaterialPath);
-
-    bool bUseDither = false;
-    Params->TryGetBoolField(TEXT("use_dither"), bUseDither);
-
-    UWorld* World = GetEditorWorld();
-    if (!World)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to get editor world"));
-    }
-
-    // Prevent duplicate proxy names
-    if (FindDraftProxyActor(World, ProxyName))
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(FString::Printf(TEXT("Draft proxy '%s' already exists. Use update_draft_proxy instead."), *ProxyName));
-    }
-
-    FScopedTransaction Transaction(FText::FromString(FString::Printf(TEXT("UnrealMCP: Create Draft Proxy %s"), *ProxyName)));
-
-    FActorSpawnParameters SpawnParams;
-    SpawnParams.Name = *FString::Printf(TEXT("DraftProxy_%s"), *ProxyName);
-    AActor* ProxyActor = World->SpawnActor<AActor>(AActor::StaticClass(), FVector::ZeroVector, FRotator::ZeroRotator, SpawnParams);
-    if (!ProxyActor)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to spawn draft proxy actor"));
-    }
-
-    ProxyActor->SetActorLabel(*ProxyName);
-    ProxyActor->Tags.AddUnique(FName(TEXT("managed_by_mcp")));
-    ProxyActor->Tags.AddUnique(FName(*FString::Printf(TEXT("mcp_id:draft_%s"), *ProxyName)));
-
-    UHierarchicalInstancedStaticMeshComponent* HISM = GetOrCreateHismComponent(ProxyActor);
-    if (!HISM)
-    {
-        ProxyActor->Destroy();
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to create HISM component"));
-    }
-
-    // Load static mesh
-    UStaticMesh* Mesh = Cast<UStaticMesh>(UEditorAssetLibrary::LoadAsset(MeshPath));
-    if (!Mesh)
-    {
-        ProxyActor->Destroy();
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(FString::Printf(TEXT("Failed to load static mesh: %s"), *MeshPath));
-    }
-    FlushRenderingCommands();
-    HISM->SetStaticMesh(Mesh);
-
-    // Load material if provided, otherwise create a default Unlit translucent draft material
-    UMaterialInterface* Material = nullptr;
-    if (!MaterialPath.IsEmpty())
-    {
-        Material = Cast<UMaterialInterface>(UEditorAssetLibrary::LoadAsset(MaterialPath));
-        if (!Material)
-        {
-            UE_LOG(LogTemp, Warning, TEXT("Could not load material at path: %s, falling back to default draft material"), *MaterialPath);
-        }
-    }
-    if (!Material)
-    {
-        Material = LoadDraftMaterial(ProxyActor, bUseDither);
-    }
-    if (Material)
-    {
-        HISM->SetMaterial(0, Material);
-    }
-
-    // Disable collision and shadows for draft visualization
-    HISM->SetCollisionEnabled(ECollisionEnabled::NoCollision);
-    HISM->SetCastShadow(false);
-
-    // Parse and add instances
-    const TArray<TSharedPtr<FJsonValue>>* InstancesArray = nullptr;
-    int32 InstanceCount = 0;
-    if (Params->TryGetArrayField(TEXT("instances"), InstancesArray) && InstancesArray)
-    {
-        for (const TSharedPtr<FJsonValue>& Value : *InstancesArray)
-        {
-            if (!Value.IsValid() || Value->Type != EJson::Object) continue;
-            FVector Location = FVector::ZeroVector;
-            FVector Scale = FVector::OneVector;
-            FRotator Rotation = FRotator::ZeroRotator;
-            if (ParseInstanceEntry(Value->AsObject(), Location, Scale, Rotation))
-            {
-                FTransform InstanceTransform(Rotation, Location, Scale);
-                HISM->AddInstance(InstanceTransform);
-                InstanceCount++;
-            }
-        }
-    }
-
-    HISM->MarkRenderStateDirty();
-
-    FEpicUnrealMCPCommonUtils::GetActorIndex().AddActor(ProxyActor);
-
-    TSharedPtr<FJsonObject> ResultObj = MakeShared<FJsonObject>();
-    ResultObj->SetBoolField(TEXT("success"), true);
-    ResultObj->SetStringField(TEXT("proxy_name"), ProxyName);
-    ResultObj->SetStringField(TEXT("actor_name"), ProxyActor->GetName());
-    ResultObj->SetNumberField(TEXT("instance_count"), InstanceCount);
-    return ResultObj;
-}
-
-TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleUpdateDraftProxy(const TSharedPtr<FJsonObject>& Params)
-{
-    FString ProxyName;
-    if (!Params->TryGetStringField(TEXT("proxy_name"), ProxyName) || ProxyName.IsEmpty())
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Missing 'proxy_name' parameter"));
-    }
-
-    UWorld* World = GetEditorWorld();
-    if (!World)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to get editor world"));
-    }
-
-    AActor* ProxyActor = FindDraftProxyActor(World, ProxyName);
-    if (!ProxyActor)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(FString::Printf(TEXT("Draft proxy '%s' not found"), *ProxyName));
-    }
-
-    UHierarchicalInstancedStaticMeshComponent* HISM = GetOrCreateHismComponent(ProxyActor);
-    if (!HISM)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to get HISM component"));
-    }
-
-    // Clear existing instances and re-add
-    FlushRenderingCommands();
-    HISM->ClearInstances();
-
-    FString MaterialPath;
-    Params->TryGetStringField(TEXT("material_path"), MaterialPath);
-    bool bUseDither = false;
-    Params->TryGetBoolField(TEXT("use_dither"), bUseDither);
-
-    UMaterialInterface* Material = nullptr;
-    if (!MaterialPath.IsEmpty())
-    {
-        Material = Cast<UMaterialInterface>(UEditorAssetLibrary::LoadAsset(MaterialPath));
-    }
-    if (!Material)
-    {
-        Material = LoadDraftMaterial(ProxyActor, bUseDither);
-    }
-    if (Material)
-    {
-        HISM->SetMaterial(0, Material);
-    }
-
-    const TArray<TSharedPtr<FJsonValue>>* InstancesArray = nullptr;
-    int32 InstanceCount = 0;
-    if (Params->TryGetArrayField(TEXT("instances"), InstancesArray) && InstancesArray)
-    {
-        for (const TSharedPtr<FJsonValue>& Value : *InstancesArray)
-        {
-            if (!Value.IsValid() || Value->Type != EJson::Object) continue;
-            FVector Location = FVector::ZeroVector;
-            FVector Scale = FVector::OneVector;
-            FRotator Rotation = FRotator::ZeroRotator;
-            if (ParseInstanceEntry(Value->AsObject(), Location, Scale, Rotation))
-            {
-                FTransform InstanceTransform(Rotation, Location, Scale);
-                HISM->AddInstance(InstanceTransform);
-                InstanceCount++;
-            }
-        }
-    }
-
-    HISM->MarkRenderStateDirty();
-
-    TSharedPtr<FJsonObject> ResultObj = MakeShared<FJsonObject>();
-    ResultObj->SetBoolField(TEXT("success"), true);
-    ResultObj->SetStringField(TEXT("proxy_name"), ProxyName);
-    ResultObj->SetNumberField(TEXT("instance_count"), InstanceCount);
-    return ResultObj;
-}
-
-TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleDeleteDraftProxy(const TSharedPtr<FJsonObject>& Params)
-{
-    FString ProxyName;
-    if (!Params->TryGetStringField(TEXT("proxy_name"), ProxyName) || ProxyName.IsEmpty())
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Missing 'proxy_name' parameter"));
-    }
-
-    UWorld* World = GetEditorWorld();
-    if (!World)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to get editor world"));
-    }
-
-    AActor* ProxyActor = FindDraftProxyActor(World, ProxyName);
-    if (!ProxyActor)
-    {
-        TSharedPtr<FJsonObject> ResultObj = MakeShared<FJsonObject>();
-        ResultObj->SetBoolField(TEXT("success"), true);
-        ResultObj->SetBoolField(TEXT("deleted"), false);
-        ResultObj->SetStringField(TEXT("message"), FString::Printf(TEXT("Draft proxy '%s' not found (already deleted)"), *ProxyName));
-        return ResultObj;
-    }
-
-    FEpicUnrealMCPCommonUtils::GetActorIndex().RemoveActor(ProxyActor);
-
-    FScopedTransaction Transaction(FText::FromString(FString::Printf(TEXT("UnrealMCP: Delete Draft Proxy %s"), *ProxyName)));
-    ProxyActor->Modify();
-    ProxyActor->Destroy();
-
-    TSharedPtr<FJsonObject> ResultObj = MakeShared<FJsonObject>();
-    ResultObj->SetBoolField(TEXT("success"), true);
-    ResultObj->SetBoolField(TEXT("deleted"), true);
-    ResultObj->SetStringField(TEXT("proxy_name"), ProxyName);
-    return ResultObj;
-}
-
-// ------------------------------------------------------------------
-// InstanceSet commands (HISM/ISM bulk instancing)
-// ------------------------------------------------------------------
-
-static AActor* FindInstanceSetActor(UWorld* World, const FString& SetId)
-{
-    if (!World) return nullptr;
-    FString McpIdTag = FString::Printf(TEXT("mcp_id:instance_set_%s"), *SetId);
-    for (TActorIterator<AActor> It(World); It; ++It)
-    {
-        if (It->Tags.Contains(FName(*McpIdTag)))
-        {
-            return *It;
-        }
-    }
-    return nullptr;
-}
-
-static UInstancedStaticMeshComponent* GetOrCreateIsmComponent(AActor* Actor, bool bUseHism)
-{
-    if (!Actor) return nullptr;
-    if (bUseHism)
-    {
-        UHierarchicalInstancedStaticMeshComponent* HISM = Actor->FindComponentByClass<UHierarchicalInstancedStaticMeshComponent>();
-        if (!HISM)
-        {
-            HISM = NewObject<UHierarchicalInstancedStaticMeshComponent>(Actor, UHierarchicalInstancedStaticMeshComponent::StaticClass(), TEXT("InstanceSetHISM"));
-            HISM->RegisterComponent();
-            if (Actor->GetRootComponent())
-            {
-                HISM->AttachToComponent(Actor->GetRootComponent(), FAttachmentTransformRules::KeepRelativeTransform);
-            }
-            else
-            {
-                Actor->SetRootComponent(HISM);
-            }
-        }
-        return HISM;
-    }
-    else
-    {
-        UInstancedStaticMeshComponent* ISM = Actor->FindComponentByClass<UInstancedStaticMeshComponent>();
-        if (!ISM)
-        {
-            ISM = NewObject<UInstancedStaticMeshComponent>(Actor, UInstancedStaticMeshComponent::StaticClass(), TEXT("InstanceSetISM"));
-            ISM->RegisterComponent();
-            if (Actor->GetRootComponent())
-            {
-                ISM->AttachToComponent(Actor->GetRootComponent(), FAttachmentTransformRules::KeepRelativeTransform);
-            }
-            else
-            {
-                Actor->SetRootComponent(ISM);
-            }
-        }
-        return ISM;
-    }
-}
-
-static bool ParseTransformArrayEntry(const TSharedPtr<FJsonObject>& Obj, FVector& OutLocation, FVector& OutScale, FRotator& OutRotation)
-{
-    if (!Obj.IsValid()) return false;
-
-    auto GetArrayDouble = [](const TArray<TSharedPtr<FJsonValue>>& Arr, int32 Index, double Default) -> double {
-        if (!Arr.IsValidIndex(Index)) return Default;
-        return Arr[Index]->AsNumber();
-    };
-
-    const TArray<TSharedPtr<FJsonValue>>* LocArray = nullptr;
-    if (Obj->TryGetArrayField(TEXT("location"), LocArray) && LocArray && LocArray->Num() >= 3)
-    {
-        OutLocation = FVector(
-            GetArrayDouble(*LocArray, 0, 0.0),
-            GetArrayDouble(*LocArray, 1, 0.0),
-            GetArrayDouble(*LocArray, 2, 0.0)
-        );
-    }
-    else
-    {
-        OutLocation = FVector::ZeroVector;
-    }
-
-    const TArray<TSharedPtr<FJsonValue>>* ScaleArray = nullptr;
-    if (Obj->TryGetArrayField(TEXT("scale"), ScaleArray) && ScaleArray && ScaleArray->Num() >= 3)
-    {
-        OutScale = FVector(
-            GetArrayDouble(*ScaleArray, 0, 1.0),
-            GetArrayDouble(*ScaleArray, 1, 1.0),
-            GetArrayDouble(*ScaleArray, 2, 1.0)
-        );
-    }
-    else
-    {
-        OutScale = FVector::OneVector;
-    }
-
-    const TArray<TSharedPtr<FJsonValue>>* RotArray = nullptr;
-    if (Obj->TryGetArrayField(TEXT("rotation"), RotArray) && RotArray && RotArray->Num() >= 3)
-    {
-        OutRotation = FRotator(
-            GetArrayDouble(*RotArray, 0, 0.0),
-            GetArrayDouble(*RotArray, 1, 0.0),
-            GetArrayDouble(*RotArray, 2, 0.0)
-        );
-    }
-    else
-    {
-        OutRotation = FRotator::ZeroRotator;
-    }
-
-    return true;
-}
-
-TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleSpawnInstanceSet(const TSharedPtr<FJsonObject>& Params)
-{
-    FString SetId;
-    if (!Params->TryGetStringField(TEXT("set_id"), SetId) || SetId.IsEmpty())
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Missing 'set_id' parameter"));
-    }
-
-    FString MeshPath;
-    if (!Params->TryGetStringField(TEXT("mesh_path"), MeshPath) || MeshPath.IsEmpty())
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Missing 'mesh_path' parameter"));
-    }
-
-    FString MaterialPath;
-    Params->TryGetStringField(TEXT("material_path"), MaterialPath);
-
-    UWorld* World = GetEditorWorld();
-    if (!World)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to get editor world"));
-    }
-
-    if (FindInstanceSetActor(World, SetId))
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(FString::Printf(TEXT("Instance set '%s' already exists. Use update_instance_set instead."), *SetId));
-    }
-
-    const TArray<TSharedPtr<FJsonValue>>* TransformsArray = nullptr;
-    int32 InstanceCount = 0;
-    if (Params->TryGetArrayField(TEXT("transforms"), TransformsArray) && TransformsArray)
-    {
-        InstanceCount = TransformsArray->Num();
-    }
-
-    bool bUseHism = InstanceCount > 100;
-
-    FScopedTransaction Transaction(FText::FromString(FString::Printf(TEXT("UnrealMCP: Spawn Instance Set %s"), *SetId)));
-
-    FActorSpawnParameters SpawnParams;
-    SpawnParams.Name = *FString::Printf(TEXT("InstanceSet_%s"), *SetId);
-    AActor* SetActor = World->SpawnActor<AActor>(AActor::StaticClass(), FVector::ZeroVector, FRotator::ZeroRotator, SpawnParams);
-    if (!SetActor)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to spawn instance set actor"));
-    }
-
-    SetActor->SetActorLabel(*SetId);
-    SetActor->Tags.AddUnique(FName(TEXT("managed_by_mcp")));
-    SetActor->Tags.AddUnique(FName(*FString::Printf(TEXT("mcp_id:instance_set_%s"), *SetId)));
-
-    UInstancedStaticMeshComponent* ISM = GetOrCreateIsmComponent(SetActor, bUseHism);
-    if (!ISM)
-    {
-        SetActor->Destroy();
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to create ISM/HISM component"));
-    }
-
-    UStaticMesh* Mesh = Cast<UStaticMesh>(UEditorAssetLibrary::LoadAsset(MeshPath));
-    if (!Mesh)
-    {
-        SetActor->Destroy();
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(FString::Printf(TEXT("Failed to load static mesh: %s"), *MeshPath));
-    }
-    FlushRenderingCommands();
-    ISM->SetStaticMesh(Mesh);
-
-    if (!MaterialPath.IsEmpty())
-    {
-        UMaterialInterface* Material = Cast<UMaterialInterface>(UEditorAssetLibrary::LoadAsset(MaterialPath));
-        if (Material)
-        {
-            ISM->SetMaterial(0, Material);
-        }
-    }
-
-    if (TransformsArray)
-    {
-        for (const TSharedPtr<FJsonValue>& Value : *TransformsArray)
-        {
-            if (!Value.IsValid() || Value->Type != EJson::Object) continue;
-            FVector Location = FVector::ZeroVector;
-            FVector Scale = FVector::OneVector;
-            FRotator Rotation = FRotator::ZeroRotator;
-            if (ParseTransformArrayEntry(Value->AsObject(), Location, Scale, Rotation))
-            {
-                FTransform InstanceTransform(Rotation, Location, Scale);
-                ISM->AddInstance(InstanceTransform);
-            }
-        }
-    }
-
-    ISM->MarkRenderStateDirty();
-    FEpicUnrealMCPCommonUtils::GetActorIndex().AddActor(SetActor);
-
-    TSharedPtr<FJsonObject> ResultObj = MakeShared<FJsonObject>();
-    ResultObj->SetBoolField(TEXT("success"), true);
-    ResultObj->SetStringField(TEXT("set_id"), SetId);
-    ResultObj->SetStringField(TEXT("actor_name"), SetActor->GetName());
-    ResultObj->SetNumberField(TEXT("instance_count"), InstanceCount);
-    ResultObj->SetBoolField(TEXT("use_hism"), bUseHism);
-    return ResultObj;
-}
-
-TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleUpdateInstanceSet(const TSharedPtr<FJsonObject>& Params)
-{
-    FString SetId;
-    if (!Params->TryGetStringField(TEXT("set_id"), SetId) || SetId.IsEmpty())
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Missing 'set_id' parameter"));
-    }
-
-    UWorld* World = GetEditorWorld();
-    if (!World)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to get editor world"));
-    }
-
-    AActor* SetActor = FindInstanceSetActor(World, SetId);
-    if (!SetActor)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(FString::Printf(TEXT("Instance set '%s' not found"), *SetId));
-    }
-
-    UInstancedStaticMeshComponent* ISM = SetActor->FindComponentByClass<UInstancedStaticMeshComponent>();
-    if (!ISM)
-    {
-        ISM = SetActor->FindComponentByClass<UHierarchicalInstancedStaticMeshComponent>();
-    }
-    if (!ISM)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to find ISM/HISM component on instance set actor"));
-    }
-
-    FlushRenderingCommands();
-    ISM->ClearInstances();
-
-    FString MaterialPath;
-    Params->TryGetStringField(TEXT("material_path"), MaterialPath);
-    if (!MaterialPath.IsEmpty())
-    {
-        UMaterialInterface* Material = Cast<UMaterialInterface>(UEditorAssetLibrary::LoadAsset(MaterialPath));
-        if (Material)
-        {
-            ISM->SetMaterial(0, Material);
-        }
-    }
-
-    const TArray<TSharedPtr<FJsonValue>>* TransformsArray = nullptr;
-    int32 InstanceCount = 0;
-    if (Params->TryGetArrayField(TEXT("transforms"), TransformsArray) && TransformsArray)
-    {
-        for (const TSharedPtr<FJsonValue>& Value : *TransformsArray)
-        {
-            if (!Value.IsValid() || Value->Type != EJson::Object) continue;
-            FVector Location = FVector::ZeroVector;
-            FVector Scale = FVector::OneVector;
-            FRotator Rotation = FRotator::ZeroRotator;
-            if (ParseTransformArrayEntry(Value->AsObject(), Location, Scale, Rotation))
-            {
-                FTransform InstanceTransform(Rotation, Location, Scale);
-                ISM->AddInstance(InstanceTransform);
-                InstanceCount++;
-            }
-        }
-    }
-
-    ISM->MarkRenderStateDirty();
-
-    TSharedPtr<FJsonObject> ResultObj = MakeShared<FJsonObject>();
-    ResultObj->SetBoolField(TEXT("success"), true);
-    ResultObj->SetStringField(TEXT("set_id"), SetId);
-    ResultObj->SetNumberField(TEXT("instance_count"), InstanceCount);
-    return ResultObj;
-}
-
-TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleDeleteInstanceSet(const TSharedPtr<FJsonObject>& Params)
-{
-    FString SetId;
-    if (!Params->TryGetStringField(TEXT("set_id"), SetId) || SetId.IsEmpty())
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Missing 'set_id' parameter"));
-    }
-
-    UWorld* World = GetEditorWorld();
-    if (!World)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to get editor world"));
-    }
-
-    AActor* SetActor = FindInstanceSetActor(World, SetId);
-    if (!SetActor)
-    {
-        TSharedPtr<FJsonObject> ResultObj = MakeShared<FJsonObject>();
-        ResultObj->SetBoolField(TEXT("success"), true);
-        ResultObj->SetBoolField(TEXT("deleted"), false);
-        ResultObj->SetStringField(TEXT("message"), FString::Printf(TEXT("Instance set '%s' not found (already deleted)"), *SetId));
-        return ResultObj;
-    }
-
-    FEpicUnrealMCPCommonUtils::GetActorIndex().RemoveActor(SetActor);
-
-    FScopedTransaction Transaction(FText::FromString(FString::Printf(TEXT("UnrealMCP: Delete Instance Set %s"), *SetId)));
-    SetActor->Modify();
-    SetActor->Destroy();
-
-    TSharedPtr<FJsonObject> ResultObj = MakeShared<FJsonObject>();
-    ResultObj->SetBoolField(TEXT("success"), true);
-    ResultObj->SetBoolField(TEXT("deleted"), true);
-    ResultObj->SetStringField(TEXT("set_id"), SetId);
-    return ResultObj;
-}
-
-TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleGetInstanceSetState(const TSharedPtr<FJsonObject>& Params)
-{
-    FString SetId;
-    if (!Params->TryGetStringField(TEXT("set_id"), SetId) || SetId.IsEmpty())
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Missing 'set_id' parameter"));
-    }
-
-    UWorld* World = GetEditorWorld();
-    if (!World)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to get editor world"));
-    }
-
-    AActor* SetActor = FindInstanceSetActor(World, SetId);
-    if (!SetActor)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(FString::Printf(TEXT("Instance set '%s' not found"), *SetId));
-    }
-
-    UInstancedStaticMeshComponent* ISM = SetActor->FindComponentByClass<UInstancedStaticMeshComponent>();
-    bool bIsHism = false;
-    if (!ISM)
-    {
-        ISM = SetActor->FindComponentByClass<UHierarchicalInstancedStaticMeshComponent>();
-        bIsHism = (ISM != nullptr);
-    }
-
-    int32 InstanceCount = ISM ? ISM->GetInstanceCount() : 0;
-    FString MeshPath;
-    FString MaterialPath;
-    if (ISM && ISM->GetStaticMesh())
-    {
-        MeshPath = ISM->GetStaticMesh()->GetPathName();
-    }
-    if (ISM && ISM->GetMaterial(0))
-    {
-        MaterialPath = ISM->GetMaterial(0)->GetPathName();
-    }
-
-    TSharedPtr<FJsonObject> ResultObj = MakeShared<FJsonObject>();
-    ResultObj->SetBoolField(TEXT("success"), true);
-    ResultObj->SetStringField(TEXT("set_id"), SetId);
-    ResultObj->SetStringField(TEXT("actor_name"), SetActor->GetName());
-    ResultObj->SetNumberField(TEXT("instance_count"), InstanceCount);
-    ResultObj->SetBoolField(TEXT("use_hism"), bIsHism);
-    if (!MeshPath.IsEmpty())
-    {
-        ResultObj->SetStringField(TEXT("mesh_path"), MeshPath);
-    }
-    if (!MaterialPath.IsEmpty())
-    {
-        ResultObj->SetStringField(TEXT("material_path"), MaterialPath);
-    }
-
-    TSharedPtr<FJsonObject> StateObj = MakeShared<FJsonObject>();
-    StateObj->SetStringField(TEXT("set_id"), SetId);
-    StateObj->SetStringField(TEXT("actor_name"), SetActor->GetName());
-    StateObj->SetNumberField(TEXT("instance_count"), InstanceCount);
-    StateObj->SetBoolField(TEXT("use_hism"), bIsHism);
-    if (!MeshPath.IsEmpty())
-    {
-        StateObj->SetStringField(TEXT("mesh_path"), MeshPath);
-    }
-    if (!MaterialPath.IsEmpty())
-    {
-        StateObj->SetStringField(TEXT("material_path"), MaterialPath);
-    }
-    ResultObj->SetObjectField(TEXT("state"), StateObj);
-    return ResultObj;
-}
-
-TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleListInstanceSets(const TSharedPtr<FJsonObject>& Params)
-{
-    UWorld* World = GetEditorWorld();
-    if (!World)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to get editor world"));
-    }
-
-    TArray<TSharedPtr<FJsonValue>> SetArray;
-    for (TActorIterator<AActor> It(World); It; ++It)
-    {
-        AActor* Actor = *It;
-        if (!Actor || !Actor->Tags.Contains(FName(TEXT("managed_by_mcp"))))
-        {
-            continue;
-        }
-        FString SetId;
-        for (const FName& Tag : Actor->Tags)
-        {
-            FString TagStr = Tag.ToString();
-            if (TagStr.StartsWith(TEXT("mcp_id:instance_set_")))
-            {
-                SetId = TagStr.RightChop(FCString::Strlen(TEXT("mcp_id:instance_set_")));
-                break;
-            }
-        }
-        if (SetId.IsEmpty())
-        {
-            continue;
-        }
-
-        UInstancedStaticMeshComponent* ISM = Actor->FindComponentByClass<UInstancedStaticMeshComponent>();
-        bool bIsHism = false;
-        if (!ISM)
-        {
-            ISM = Actor->FindComponentByClass<UHierarchicalInstancedStaticMeshComponent>();
-            bIsHism = (ISM != nullptr);
-        }
-
-        int32 InstanceCount = ISM ? ISM->GetInstanceCount() : 0;
-        FString MeshPath;
-        FString MaterialPath;
-        if (ISM && ISM->GetStaticMesh())
-        {
-            MeshPath = ISM->GetStaticMesh()->GetPathName();
-        }
-        if (ISM && ISM->GetMaterial(0))
-        {
-            MaterialPath = ISM->GetMaterial(0)->GetPathName();
-        }
-
-        TSharedPtr<FJsonObject> SetObj = MakeShared<FJsonObject>();
-        SetObj->SetStringField(TEXT("set_id"), SetId);
-        SetObj->SetStringField(TEXT("actor_name"), Actor->GetName());
-        SetObj->SetNumberField(TEXT("instance_count"), InstanceCount);
-        SetObj->SetBoolField(TEXT("use_hism"), bIsHism);
-        if (!MeshPath.IsEmpty())
-        {
-            SetObj->SetStringField(TEXT("mesh_path"), MeshPath);
-        }
-        if (!MaterialPath.IsEmpty())
-        {
-            SetObj->SetStringField(TEXT("material_path"), MaterialPath);
-        }
-        SetArray.Add(MakeShared<FJsonValueObject>(SetObj));
-    }
-
-    TSharedPtr<FJsonObject> ResultObj = MakeShared<FJsonObject>();
-    ResultObj->SetBoolField(TEXT("success"), true);
-    ResultObj->SetNumberField(TEXT("count"), SetArray.Num());
-    ResultObj->SetArrayField(TEXT("sets"), SetArray);
-    return ResultObj;
-}
-
-// ------------------------------------------------------------------
-// Physics commands
-// ------------------------------------------------------------------
-
-TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleSetActorCollisionPreset(const TSharedPtr<FJsonObject>& Params)
-{
-    FString ActorName;
-    if (!Params->TryGetStringField(TEXT("actor_name"), ActorName) || ActorName.IsEmpty())
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Missing or empty 'actor_name' parameter"));
-    }
-
-    FString PresetName;
-    if (!Params->TryGetStringField(TEXT("preset"), PresetName) || PresetName.IsEmpty())
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Missing or empty 'preset' parameter"));
-    }
-
-    UWorld* World = GetEditorWorld();
-    if (!World)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to get editor world"));
-    }
-
-    AActor* Actor = nullptr;
-    for (TActorIterator<AActor> It(World); It; ++It)
-    {
-        if (It->GetName() == ActorName || It->GetActorLabel() == ActorName)
-        {
-            Actor = *It;
-            break;
-        }
-    }
-
-    if (!Actor)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(FString::Printf(TEXT("Actor '%s' not found"), *ActorName));
-    }
-
-    UPrimitiveComponent* RootComp = Cast<UPrimitiveComponent>(Actor->GetRootComponent());
-    if (!RootComp)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Actor does not have a primitive root component"));
-    }
-
-    FScopedTransaction Transaction(FText::FromString(FString::Printf(TEXT("UnrealMCP: Set Collision Preset %s"), *ActorName)));
-    Actor->Modify();
-    RootComp->Modify();
-    RootComp->SetCollisionProfileName(FName(*PresetName));
-
-    TSharedPtr<FJsonObject> Result = MakeShared<FJsonObject>();
-    Result->SetBoolField(TEXT("success"), true);
-    Result->SetStringField(TEXT("actor_name"), ActorName);
-    Result->SetStringField(TEXT("preset"), PresetName);
-    return Result;
-}
-
-TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleSetActorPhysics(const TSharedPtr<FJsonObject>& Params)
-{
-    FString ActorName;
-    if (!Params->TryGetStringField(TEXT("actor_name"), ActorName) || ActorName.IsEmpty())
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Missing or empty 'actor_name' parameter"));
-    }
-
-    UWorld* World = GetEditorWorld();
-    if (!World)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to get editor world"));
-    }
-
-    AActor* Actor = nullptr;
-    for (TActorIterator<AActor> It(World); It; ++It)
-    {
-        if (It->GetName() == ActorName || It->GetActorLabel() == ActorName)
-        {
-            Actor = *It;
-            break;
-        }
-    }
-
-    if (!Actor)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(FString::Printf(TEXT("Actor '%s' not found"), *ActorName));
-    }
-
-    UPrimitiveComponent* RootComp = Cast<UPrimitiveComponent>(Actor->GetRootComponent());
-    if (!RootComp)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Actor does not have a primitive root component"));
-    }
-
-    FScopedTransaction Transaction(FText::FromString(FString::Printf(TEXT("UnrealMCP: Set Actor Physics %s"), *ActorName)));
-    Actor->Modify();
-    RootComp->Modify();
-
-    bool bSimulatePhysics = false;
-    if (Params->TryGetBoolField(TEXT("simulate_physics"), bSimulatePhysics))
-    {
-        RootComp->SetSimulatePhysics(bSimulatePhysics);
-    }
-
-    bool bGravityEnabled = true;
-    if (Params->TryGetBoolField(TEXT("gravity_enabled"), bGravityEnabled))
-    {
-        RootComp->SetEnableGravity(bGravityEnabled);
-    }
-
-    double MassScale = 1.0;
-    if (Params->TryGetNumberField(TEXT("mass_scale"), MassScale))
-    {
-        RootComp->SetMassScale(NAME_None, static_cast<float>(MassScale));
-    }
-
-    double LinearDamping = 0.01;
-    if (Params->TryGetNumberField(TEXT("linear_damping"), LinearDamping))
-    {
-        RootComp->SetLinearDamping(static_cast<float>(LinearDamping));
-    }
-
-    double AngularDamping = 0.0;
-    if (Params->TryGetNumberField(TEXT("angular_damping"), AngularDamping))
-    {
-        RootComp->SetAngularDamping(static_cast<float>(AngularDamping));
-    }
-
-    TSharedPtr<FJsonObject> Result = MakeShared<FJsonObject>();
-    Result->SetBoolField(TEXT("success"), true);
-    Result->SetStringField(TEXT("actor_name"), ActorName);
-    Result->SetBoolField(TEXT("simulate_physics"), RootComp->IsSimulatingPhysics());
-    return Result;
-}
-
-TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleCreatePhysicalMaterial(const TSharedPtr<FJsonObject>& Params)
-{
-    FString AssetPath;
-    if (!Params->TryGetStringField(TEXT("asset_path"), AssetPath) || AssetPath.IsEmpty())
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Missing or empty 'asset_path' parameter"));
-    }
-
-    FString AssetName = FPaths::GetBaseFilename(AssetPath);
-    UPackage* Package = CreatePackage(*AssetPath);
-    if (!Package)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to create package for physical material"));
-    }
-
-    UPhysicalMaterial* PhysMat = NewObject<UPhysicalMaterial>(Package, FName(*AssetName), RF_Public | RF_Standalone);
-    if (!PhysMat)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to create PhysicalMaterial object"));
-    }
-
-    double Friction = 0.7;
-    if (Params->TryGetNumberField(TEXT("friction"), Friction))
-    {
-        PhysMat->Friction = static_cast<float>(Friction);
-    }
-
-    double Restitution = 0.3;
-    if (Params->TryGetNumberField(TEXT("restitution"), Restitution))
-    {
-        PhysMat->Restitution = static_cast<float>(Restitution);
-    }
-
-    double Density = 1.0;
-    if (Params->TryGetNumberField(TEXT("density"), Density))
-    {
-        // Density is not a direct property on UPhysicalMaterial in newer UE versions;
-        // it is typically part of the body setup or physical material mask.
-        // We skip setting it here to avoid compilation issues.
-    }
-
-    Package->MarkPackageDirty();
-    FAssetRegistryModule::AssetCreated(PhysMat);
-
-    TSharedPtr<FJsonObject> Result = MakeShared<FJsonObject>();
-    Result->SetBoolField(TEXT("success"), true);
-    Result->SetStringField(TEXT("asset_path"), AssetPath);
-    Result->SetStringField(TEXT("asset_name"), AssetName);
-    return Result;
-}
-
-TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleSpawnRadialForce(const TSharedPtr<FJsonObject>& Params)
-{
-    FString ActorName = TEXT("RadialForceActor");
-    Params->TryGetStringField(TEXT("actor_name"), ActorName);
-
-    UWorld* World = GetEditorWorld();
-    if (!World)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to get editor world"));
-    }
-
-    FVector Location = FVector::ZeroVector;
-    const TSharedPtr<FJsonObject>* LocObj = nullptr;
-    if (Params->TryGetObjectField(TEXT("location"), LocObj) && LocObj)
-    {
-        double X = 0, Y = 0, Z = 0;
-        (*LocObj)->TryGetNumberField(TEXT("x"), X);
-        (*LocObj)->TryGetNumberField(TEXT("y"), Y);
-        (*LocObj)->TryGetNumberField(TEXT("z"), Z);
-        Location = FVector(X, Y, Z);
-    }
-
-    FScopedTransaction Transaction(FText::FromString(FString::Printf(TEXT("UnrealMCP: Spawn Radial Force %s"), *ActorName)));
-
-    ARadialForceActor* ForceActor = World->SpawnActor<ARadialForceActor>(ARadialForceActor::StaticClass(), Location, FRotator::ZeroRotator);
-    if (!ForceActor)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to spawn RadialForceActor"));
-    }
-
-    ForceActor->SetActorLabel(*ActorName);
-    ForceActor->Tags.AddUnique(FName(TEXT("managed_by_mcp")));
-
-    double Radius = 500.0;
-    if (Params->TryGetNumberField(TEXT("radius"), Radius))
-    {
-        if (ForceActor->GetForceComponent())
-        {
-            ForceActor->GetForceComponent()->Radius = static_cast<float>(Radius);
-        }
-    }
-
-    double Strength = 1000.0;
-    if (Params->TryGetNumberField(TEXT("strength"), Strength))
-    {
-        if (ForceActor->GetForceComponent())
-        {
-            ForceActor->GetForceComponent()->ForceStrength = static_cast<float>(Strength);
-        }
-    }
-
-    FEpicUnrealMCPCommonUtils::GetActorIndex().AddActor(ForceActor);
-
-    TSharedPtr<FJsonObject> Result = MakeShared<FJsonObject>();
-    Result->SetBoolField(TEXT("success"), true);
-    Result->SetStringField(TEXT("actor_name"), ForceActor->GetName());
-    Result->SetNumberField(TEXT("radius"), Radius);
-    Result->SetNumberField(TEXT("strength"), Strength);
-    return Result;
-}
-
-TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleSpawnPhysicsConstraint(const TSharedPtr<FJsonObject>& Params)
-{
-    FString ActorName = TEXT("PhysicsConstraintActor");
-    Params->TryGetStringField(TEXT("actor_name"), ActorName);
-
-    UWorld* World = GetEditorWorld();
-    if (!World)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to get editor world"));
-    }
-
-    FVector Location = FVector::ZeroVector;
-    const TSharedPtr<FJsonObject>* LocObj = nullptr;
-    if (Params->TryGetObjectField(TEXT("location"), LocObj) && LocObj)
-    {
-        double X = 0, Y = 0, Z = 0;
-        (*LocObj)->TryGetNumberField(TEXT("x"), X);
-        (*LocObj)->TryGetNumberField(TEXT("y"), Y);
-        (*LocObj)->TryGetNumberField(TEXT("z"), Z);
-        Location = FVector(X, Y, Z);
-    }
-
-    FScopedTransaction Transaction(FText::FromString(FString::Printf(TEXT("UnrealMCP: Spawn Physics Constraint %s"), *ActorName)));
-
-    APhysicsConstraintActor* ConstraintActor = World->SpawnActor<APhysicsConstraintActor>(APhysicsConstraintActor::StaticClass(), Location, FRotator::ZeroRotator);
-    if (!ConstraintActor)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to spawn PhysicsConstraintActor"));
-    }
-
-    ConstraintActor->SetActorLabel(*ActorName);
-    ConstraintActor->Tags.AddUnique(FName(TEXT("managed_by_mcp")));
-
-    FEpicUnrealMCPCommonUtils::GetActorIndex().AddActor(ConstraintActor);
-
-    TSharedPtr<FJsonObject> Result = MakeShared<FJsonObject>();
-    Result->SetBoolField(TEXT("success"), true);
-    Result->SetStringField(TEXT("actor_name"), ConstraintActor->GetName());
-    return Result;
-}
-
-// ------------------------------------------------------------------
-// Validation commands
-// ------------------------------------------------------------------
-
-TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleCompileAllBlueprints(const TSharedPtr<FJsonObject>& Params)
-{
-    TArray<FAssetData> BlueprintAssets;
-    FAssetRegistryModule& AssetRegistryModule = FModuleManager::LoadModuleChecked<FAssetRegistryModule>(TEXT("AssetRegistry"));
-    FARFilter Filter;
-    Filter.ClassPaths.Add(UBlueprint::StaticClass()->GetClassPathName());
-    Filter.bRecursivePaths = true;
-    AssetRegistryModule.Get().GetAssets(Filter, BlueprintAssets);
-
-    int32 CompiledCount = 0;
-    int32 ErrorCount = 0;
-    TArray<TSharedPtr<FJsonValue>> ErrorList;
-
-    for (const FAssetData& Asset : BlueprintAssets)
-    {
-        UBlueprint* Blueprint = Cast<UBlueprint>(UEditorAssetLibrary::LoadAsset(Asset.GetObjectPathString()));
-        if (!Blueprint)
-        {
-            continue;
-        }
-
-        FCompilerResultsLog Results;
-        Results.SetSourcePath(Asset.GetObjectPathString());
-        FKismetEditorUtilities::CompileBlueprint(Blueprint, EBlueprintCompileOptions::None, &Results);
-        CompiledCount++;
-
-        if (Results.NumErrors > 0)
-        {
-            ErrorCount++;
-            TSharedPtr<FJsonObject> ErrObj = MakeShared<FJsonObject>();
-            ErrObj->SetStringField(TEXT("asset"), Asset.GetObjectPathString());
-            ErrObj->SetNumberField(TEXT("errors"), Results.NumErrors);
-            ErrObj->SetNumberField(TEXT("warnings"), Results.NumWarnings);
-            ErrorList.Add(MakeShared<FJsonValueObject>(ErrObj));
-        }
-    }
-
-    TSharedPtr<FJsonObject> Result = MakeShared<FJsonObject>();
-    Result->SetBoolField(TEXT("success"), true);
-    Result->SetNumberField(TEXT("compiled_count"), CompiledCount);
-    Result->SetNumberField(TEXT("error_count"), ErrorCount);
-    Result->SetArrayField(TEXT("errors"), ErrorList);
-    return Result;
-}
-
-TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleRunMapCheck(const TSharedPtr<FJsonObject>& Params)
-{
-    UWorld* World = GetEditorWorld();
-    if (!World)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to get editor world"));
-    }
-
-    // Run the editor's map check
-    FMessageLog MapCheckLog("MapCheck");
-    MapCheckLog.NewPage(FText::FromString(TEXT("MCP Map Check")));
-
-    int32 ErrorCount = 0;
-    int32 WarningCount = 0;
-
-    for (TActorIterator<AActor> It(World); It; ++It)
-    {
-        AActor* Actor = *It;
-        if (!Actor)
-        {
-            continue;
-        }
-
-        // Check for actors without valid root component
-        if (!Actor->GetRootComponent())
-        {
-            WarningCount++;
-            MapCheckLog.Warning()
-                ->AddToken(FTextToken::Create(FText::FromString(Actor->GetName())))
-                ->AddToken(FTextToken::Create(FText::FromString(TEXT("has no root component"))));
-        }
-
-        // Check for overlapping static mesh actors (simplified)
-        AStaticMeshActor* SMActor = Cast<AStaticMeshActor>(Actor);
-        if (SMActor && SMActor->GetStaticMeshComponent())
-        {
-            if (!SMActor->GetStaticMeshComponent()->GetStaticMesh())
-            {
-                ErrorCount++;
-                MapCheckLog.Error()
-                    ->AddToken(FTextToken::Create(FText::FromString(Actor->GetName())))
-                    ->AddToken(FTextToken::Create(FText::FromString(TEXT("has no static mesh assigned"))));
-            }
-        }
-    }
-
-    TSharedPtr<FJsonObject> Result = MakeShared<FJsonObject>();
-    Result->SetBoolField(TEXT("success"), true);
-    Result->SetNumberField(TEXT("errors"), ErrorCount);
-    Result->SetNumberField(TEXT("warnings"), WarningCount);
-    Result->SetStringField(TEXT("message"), TEXT("Map check completed. See Unreal Editor's Map Check tab for details."));
-    return Result;
-}
-
-TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleFindBrokenReferences(const TSharedPtr<FJsonObject>& Params)
-{
-    UWorld* World = GetEditorWorld();
-    if (!World)
-    {
-        return FEpicUnrealMCPCommonUtils::CreateErrorResponse(TEXT("Failed to get editor world"));
-    }
-
-    TArray<TSharedPtr<FJsonValue>> BrokenActors;
-    int32 MissingMeshCount = 0;
-    int32 MissingMaterialCount = 0;
-
-    for (TActorIterator<AActor> It(World); It; ++It)
-    {
-        AActor* Actor = *It;
-        if (!Actor)
-        {
-            continue;
-        }
-
-        TArray<TSharedPtr<FJsonValue>> Issues;
-
-        AStaticMeshActor* SMActor = Cast<AStaticMeshActor>(Actor);
-        if (SMActor && SMActor->GetStaticMeshComponent())
-        {
-            if (!SMActor->GetStaticMeshComponent()->GetStaticMesh())
-            {
-                MissingMeshCount++;
-                TSharedPtr<FJsonObject> Issue = MakeShared<FJsonObject>();
-                Issue->SetStringField(TEXT("type"), TEXT("missing_mesh"));
-                Issue->SetStringField(TEXT("component"), TEXT("StaticMeshComponent"));
-                Issues.Add(MakeShared<FJsonValueObject>(Issue));
-            }
-
-            UMaterialInterface* Mat = SMActor->GetStaticMeshComponent()->GetMaterial(0);
-            if (!Mat)
-            {
-                MissingMaterialCount++;
-                TSharedPtr<FJsonObject> Issue = MakeShared<FJsonObject>();
-                Issue->SetStringField(TEXT("type"), TEXT("missing_material"));
-                Issue->SetStringField(TEXT("component"), TEXT("StaticMeshComponent"));
-                Issue->SetNumberField(TEXT("slot"), 0);
-                Issues.Add(MakeShared<FJsonValueObject>(Issue));
-            }
-        }
-
-        if (Issues.Num() > 0)
-        {
-            TSharedPtr<FJsonObject> ActorObj = MakeShared<FJsonObject>();
-            ActorObj->SetStringField(TEXT("actor_name"), Actor->GetName());
-            ActorObj->SetStringField(TEXT("actor_label"), Actor->GetActorLabel());
-            ActorObj->SetArrayField(TEXT("issues"), Issues);
-            BrokenActors.Add(MakeShared<FJsonValueObject>(ActorObj));
-        }
-    }
-
-    TSharedPtr<FJsonObject> Result = MakeShared<FJsonObject>();
-    Result->SetBoolField(TEXT("success"), true);
-    Result->SetNumberField(TEXT("broken_actor_count"), BrokenActors.Num());
-    Result->SetNumberField(TEXT("missing_mesh_count"), MissingMeshCount);
-    Result->SetNumberField(TEXT("missing_material_count"), MissingMaterialCount);
-    Result->SetArrayField(TEXT("broken_actors"), BrokenActors);
-    return Result;
+    return FEpicUnrealMCPCommonUtils::CreateErrorResponse(
+        FString::Printf(TEXT("Unknown procedural command: %s"), *CommandType));
 }
 
 TSharedPtr<FJsonObject> FEpicUnrealMCPProceduralCommands::HandleRequestCognitiveProcessing(const TSharedPtr<FJsonObject>& Params)
