+++ Plugins/UnrealMCP/Source/UnrealMCP/Private/Commands/EpicUnrealMCPRouter.cpp
@@ -23,22 +23,22 @@ int32 FEpicUnrealMCPRouter::RouteCommand(const FString& CommandType)
         {TEXT("create_blackboard"), 20},
         {TEXT("create_nav_modifier_volume"), 20},
         {TEXT("create_nav_link_proxy"), 20},
-        {TEXT("set_actor_collision_preset"), 19},
-        {TEXT("set_actor_physics"), 19},
-        {TEXT("create_physical_material"), 19},
-        {TEXT("spawn_radial_force"), 19},
-        {TEXT("spawn_physics_constraint"), 19},
-        {TEXT("compile_all_blueprints"), 19},
-        {TEXT("run_map_check"), 19},
-        {TEXT("find_broken_references"), 19},
-        {TEXT("create_draft_proxy"), 19},
-        {TEXT("update_draft_proxy"), 19},
-        {TEXT("delete_draft_proxy"), 19},
-        {TEXT("spawn_instance_set"), 19},
-        {TEXT("update_instance_set"), 19},
-        {TEXT("delete_instance_set"), 19},
-        {TEXT("get_instance_set_state"), 19},
-        {TEXT("list_instance_sets"), 19},
+        {TEXT("set_actor_collision_preset"), 22},
+        {TEXT("set_actor_physics"), 22},
+        {TEXT("create_physical_material"), 22},
+        {TEXT("spawn_radial_force"), 22},
+        {TEXT("spawn_physics_constraint"), 22},
+        {TEXT("compile_all_blueprints"), 23},
+        {TEXT("run_map_check"), 23},
+        {TEXT("find_broken_references"), 23},
+        {TEXT("create_draft_proxy"), 24},
+        {TEXT("update_draft_proxy"), 24},
+        {TEXT("delete_draft_proxy"), 24},
+        {TEXT("spawn_instance_set"), 24},
+        {TEXT("update_instance_set"), 24},
+        {TEXT("delete_instance_set"), 24},
+        {TEXT("get_instance_set_state"), 24},
+        {TEXT("list_instance_sets"), 24},
         {TEXT("request_cognitive_processing"), 19},
         {TEXT("create_blueprint"), 2},
         {TEXT("add_component_to_blueprint"), 2},
