+++ Plugins/UnrealMCP/Source/UnrealMCP/Public/Commands/EpicUnrealMCPProceduralCommands.h
@@ -4,9 +4,17 @@
 #include "Json.h"
 
 /**
- * Handler class for procedural generation and instancing commands.
- * Split from FEpicUnrealMCPEditorCommands to reduce file size and
- * isolate procedural/instancing logic.
+ * Handler class for procedural generation commands (route 19).
+ *
+ * Phase 4 (Issue #31) extracted Physics, Validation, and Draft/Instance
+ * commands into dedicated handler classes:
+ *   - FEpicUnrealMCPPhysicsCommands     (route 22)
+ *   - FEpicUnrealMCPValidationCommands  (route 23)
+ *   - FEpicUnrealMCPInstanceCommands    (route 24)
+ *
+ * What remains here is the procedural-generation surface itself plus
+ * request_cognitive_processing (a single command that does not yet
+ * justify its own handler class).
  */
 class UNREALMCP_API FEpicUnrealMCPProceduralCommands
 {
@@ -25,28 +33,6 @@ private:
     TSharedPtr<FJsonObject> HandleCreateDataLayerForGeneration(const TSharedPtr<FJsonObject>& Params);
     TSharedPtr<FJsonObject> HandleClearGeneratedGroup(const TSharedPtr<FJsonObject>& Params);
 
-    // Draft proxy (HISM visualization)
-    TSharedPtr<FJsonObject> HandleCreateDraftProxy(const TSharedPtr<FJsonObject>& Params);
-    TSharedPtr<FJsonObject> HandleUpdateDraftProxy(const TSharedPtr<FJsonObject>& Params);
-    TSharedPtr<FJsonObject> HandleDeleteDraftProxy(const TSharedPtr<FJsonObject>& Params);
-
-    // InstanceSet commands (HISM/ISM bulk instancing)
-    TSharedPtr<FJsonObject> HandleSpawnInstanceSet(const TSharedPtr<FJsonObject>& Params);
-    TSharedPtr<FJsonObject> HandleUpdateInstanceSet(const TSharedPtr<FJsonObject>& Params);
-    TSharedPtr<FJsonObject> HandleDeleteInstanceSet(const TSharedPtr<FJsonObject>& Params);
-    TSharedPtr<FJsonObject> HandleGetInstanceSetState(const TSharedPtr<FJsonObject>& Params);
-    TSharedPtr<FJsonObject> HandleListInstanceSets(const TSharedPtr<FJsonObject>& Params);
+    // Cognitive processing (single-command surface kept here for now)
     TSharedPtr<FJsonObject> HandleRequestCognitiveProcessing(const TSharedPtr<FJsonObject>& Params);
-
-    // Physics commands (collision, physics body, materials, forces, constraints)
-    TSharedPtr<FJsonObject> HandleSetActorCollisionPreset(const TSharedPtr<FJsonObject>& Params);
-    TSharedPtr<FJsonObject> HandleSetActorPhysics(const TSharedPtr<FJsonObject>& Params);
-    TSharedPtr<FJsonObject> HandleCreatePhysicalMaterial(const TSharedPtr<FJsonObject>& Params);
-    TSharedPtr<FJsonObject> HandleSpawnRadialForce(const TSharedPtr<FJsonObject>& Params);
-    TSharedPtr<FJsonObject> HandleSpawnPhysicsConstraint(const TSharedPtr<FJsonObject>& Params);
-
-    // Validation commands
-    TSharedPtr<FJsonObject> HandleCompileAllBlueprints(const TSharedPtr<FJsonObject>& Params);
-    TSharedPtr<FJsonObject> HandleRunMapCheck(const TSharedPtr<FJsonObject>& Params);
-    TSharedPtr<FJsonObject> HandleFindBrokenReferences(const TSharedPtr<FJsonObject>& Params);
 };
